refactor(theme): catalog traits on Theme via ThemeExt
